chore(portfolio-contract): recover from merge/rebase
 - finish factoring out makeTrader(), deploy()
